pyproject.toml
[tool.poetry]
name = "poetry_conta_backend"
version = "0.1.0"
description = ""
authors = ["Mamy-cloud <mamycompteprofessionnel@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
cryptography = ">=41.0.0,<46.0.0"
python-jose = {version = ">=3.5.0,<4.0.0", extras = ["cryptography"]}
supabase = "^2.28.3"
uvicorn = {extras = ["standard"], version = "^0.46.0"}
fastapi = "^0.136.0"
SQLAlchemy = "^2.0.49"
psycopg2-binary = "^2.9.12"
python-dotenv = "^1.2.2"
python-multipart = "^0.0.26"
passlib = "^1.7.4"
bcrypt = "^5.0.0"
pydantic-settings = "^2.14.0"
pydantic = {extras = ["email"], version = "^2.13.3"}
APScheduler = "^3.11.2"
httpx = "^0.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
